refactor(bayn): modularize cycle runner
